requirements.txt
appnope==0.1.3
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
automium==0.2.6
automium-web==0.1.1
Babel==2.10.3
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
blinker==1.5
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cors==1.0.1
cssselect==1.1.0
debugpy==1.6.2
decorator==5.1.1
Deprecated==1.2.13
dnspython==2.2.1
dominate==2.7.0
email-validator==1.2.1
entrypoints==0.4
exceptiongroup==1.1.0
executing==0.9.1
filelock==3.8.0
Flask==2.2.2
Flask-Babel==2.0.0
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Cors==3.0.10
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.5.1
Flask-Table==0.5.0
Flask-WTF==1.0.1
future==0.18.2
gevent==22.10.2
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.57.0
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
httplib2==0.21.0
idna==3.3
imageio==2.22.2
ipykernel==6.15.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.3.4
jupyter-core==4.11.1
legacy==0.1.6
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mongoengine==0.24.2
motor==3.0.0
nest-asyncio==1.5.5
oauthlib==3.2.2
outcome==1.2.0
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.6.2
prompt-toolkit==3.0.30
protobuf==4.21.12
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
PyJWT==2.4.0
pymongo==4.2.0
pyparsing==3.0.9
pyperclip==1.8.2
pyquery==1.4.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
pyzmq==23.2.0
redis==4.3.4
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spotipy==2.20.0
stack-data==0.3.0
tldextract==3.4.0
tornado==6.2
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
twilio==7.12.1
uritemplate==4.1.1
urljoin==1.0.0
urllib3==1.26.12
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
WTForms==3.0.1
zope.event==4.5.0
zope.interface==5.5.2

pip~=23.1.2
argparse~=1.4.0
wheel~=0.40.0
setuptools~=65.5.0
fpdf~=1.7.2